requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
openai-whisper
google-generativeai==0.5.4
ollama
spacy
scispacy
twilio==9.0.5
aioredis
redis>=5.0
apscheduler==3.10.4
scikit-learn>=1.5.0
opencv-python
mediapipe
pandas
numpy>=1.26,<3
pydantic==2.7.1
python-multipart==0.0.9
requests>=2.31.0
websockets>=12.0
